UI User Feedback (SOLVED)
Solved the problem of figuring out how to let the user know when they select a cube. (VISUAL FIX)

(MY SOLUTION)

Make cube move forward when mouse hovers over it. I ran into problems with this because the cube is locked to the panels on the canvas. Scaling the cubes does not run into any issues though.

Scale cubes up using Void OnMouseEnter() and, scale cubes down on Void OnMouseExit()

During the mouse hover checks I change a mouse_on variable to true or false depending on the mouse's status (hovering or not hovering on cube)

In Void Update() I have an If and Else statement that checks if mouse_on is true or false. If true I scale the cube up

If anyone is looking at this commit use Mathf.SmoothDamp() if you need to smoothly get between 2 different floats over time. Feel free to use my code in you're own projects